add optional k/v labels to Objectives (#237)
objectives section copied verbatim from existing spec
https://github.com/OpenSLO/OpenSLO/blob/b54f77a7637f3822a12b19624c5899a038958f4f/README.md?plain=1#L263C1-L280C4,
except for the addition of the labels object.

The driving use case is being able to categorize the objectives and
associate them to processes or tools outside of the SLO itself. E.g. if
there's a goal to improve an SLO Objective, it may want to be tagged
with identifiers around that larger goal, but the existing objective may
want to remain in place for existing alerting, etc.

---------

Co-authored-by: Mateusz Hawrus <48822818+nieomylnieja@users.noreply.github.com>