Make OpenSLO spec concrete (#294)
# Problem statement

We claim that OpenSLO should be vendor agnostic, this means that spec
should be running for different vendors without any additional changes.
At least to the core, vendor specific annotations should be fine. With
this bullet point there is possibility that the same spec won't work for
two vendors although both of them clam that they support OpenSLO and
they are right.

Another thing is that recently we decided that we want to make OpenSLO
not only easy to use but also easy to implement as we want new projects
to start using this spec. Taking specification concrete and simple is
step into this direction.