chore: Make object references point at names only (#304)
Currently in the README there's an ambiguous sentence referring to some
of the object references, namely: `targetRef`, `conditionRef` (Alert
Policy) and `alertPolicyRef` (SLO), which states:

> this is the name or path]

The question that arises is: "path to what?".
Unfortunately digging in [some
history](https://github.com/OpenSLO/OpenSLO/pull/54) behind this I found
no explicit mentions or discussions around the `ref` path.
I believe this might be some legacy which furthermore is not well
specified, that's why I propose to remove it from the v1 version as it's
not clear what it even means (file path? url path? yaml path?).